refactor(runner): decompose executor.ts into functional modules
Extract helper functions from executor.ts (883 lines) into focused modules:

- executor-types.ts: Shared interfaces (ExecutionResult, ExecutionOptions, PreflightResult)
- executor-env.ts: Environment variable building and path constants
- network-logs/: Network log reading, uploading, and cleanup
- vm-setup/: SSH-based VM setup operations (scripts, storages, DNS, CA, session)

Result: executor.ts reduced to 509 lines (42% reduction), focusing on
orchestration and VM lifecycle management.

Follows the decomposition pattern from PR #1287 (RunService).

Closes #1293
