+++ turbo/apps/runner/src/lib/executor-types.ts
@@ -1,0 +1,42 @@
+/**
+ * Executor Types
+ *
+ * Shared interfaces and types for the job executor.
+ */
+
+/**
+ * Execution result
+ */
+export interface ExecutionResult {
+  exitCode: number;
+  error?: string;
+}
+
+/**
+ * Execution options for customizing job execution behavior
+ */
+export interface ExecutionOptions {
+  /**
+   * Benchmark mode for local VM performance testing without API server:
+   * - Runs prompt directly as bash command (skips run-agent.py)
+   * - Skips network log upload
+   * - Skips telemetry reporting
+   * Used by the benchmark command
+   */
+  benchmarkMode?: boolean;
+
+  /**
+   * Custom logger function for execution output.
+   * If provided, executor will use this instead of console.log.
+   * Useful for adding timestamps or custom prefixes.
+   */
+  logger?: (message: string) => void;
+}
+
+/**
+ * Preflight check result
+ */
+export interface PreflightResult {
+  success: boolean;
+  error?: string;
+}
