+++ turbo/apps/runner/src/lib/vm-setup/index.ts
@@ -1,0 +1,13 @@
+/**
+ * VM Setup Module
+ *
+ * Re-exports for VM setup operations.
+ */
+
+export {
+  uploadScripts,
+  downloadStorages,
+  restoreSessionHistory,
+  installProxyCA,
+  configureDNS,
+} from "./vm-setup.js";
